fix: left click and drag when selecting entities
