+++ unity-client/Assets/Scripts/MainScripts/DCL/Controllers/BuilderMode/BuilderInWorldController.cs
@@ -264,23 +264,23 @@ public class BuilderInWorldController : MonoBehaviour
             TryStartEnterEditMode();
     }
 
-    public DCLBuilderInWorldEntity GetEntityOnPointer()
-    {
-        RaycastHit hit;
-        UnityEngine.Ray ray = Camera.main.ScreenPointToRay(biwModeController.GetMousePosition());
-        float distanceToSelect = biwModeController.GetMaxDistanceToSelectEntities();
-
-        if (Physics.Raycast(ray, out hit, distanceToSelect, layerToRaycast))
-        {
-            string entityID = hit.collider.gameObject.name;
-
-            if (sceneToEdit.entities.ContainsKey(entityID))
-            {
-                return builderInWorldEntityHandler.GetConvertedEntity(sceneToEdit.entities[entityID]);
-            }
-        }
-        return null;
-    }
+    // public DCLBuilderInWorldEntity GetEntityOnPointer()
+    // {
+    //     RaycastHit hit;
+    //     UnityEngine.Ray ray = Camera.main.ScreenPointToRay(biwModeController.GetMousePosition());
+    //     float distanceToSelect = biwModeController.GetMaxDistanceToSelectEntities();
+    //
+    //     if (Physics.Raycast(ray, out hit, distanceToSelect, layerToRaycast))
+    //     {
+    //         string entityID = hit.collider.gameObject.name;
+    //
+    //         if (sceneToEdit.entities.ContainsKey(entityID))
+    //         {
+    //             return builderInWorldEntityHandler.GetConvertedEntity(sceneToEdit.entities[entityID]);
+    //         }
+    //     }
+    //     return null;
+    // }
 
     public VoxelEntityHit GetCloserUnselectedVoxelEntityOnPointer()
     {
